Criação de endpoints para transactions e refatoração de código e regra de negócios.

+++ clients/prisma/migrations/20250731230248_initial_config_v1/migration.sql
@@ -41,7 +41,7 @@ CREATE TABLE "BankAccount" (
     "agency" TEXT NOT NULL,
     "accountNumber" TEXT NOT NULL,
     "bankName" TEXT NOT NULL,
-    "balance" BIGINT NOT NULL DEFAULT 0,
+    "balance" INTEGER NOT NULL DEFAULT 0,
 
     CONSTRAINT "BankAccount_pkey" PRIMARY KEY ("id")
 );
